Update TelegramBots & Log4j2

+++ src/main/java/space/delusive/tversu/component/MetricsRegistrar.java
@@ -3,7 +3,7 @@ package space.delusive.tversu.component;
 public interface MetricsRegistrar {
     void registerPath(String path);
 
-    void registerUserCall(Integer userId);
+    void registerUserCall(long userId);
 
     void registerTimeConsumed(long millisecond);
 }
